# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: food
    description: Food scanning and nutrition operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /food/scan:
    post:
      operationId: scanFood
      tags: [food]
      summary: Scan food image and get nutritional breakdown
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanFoodRequest"
      responses:
        "200":
          description: Food scan results with ingredient breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanFoodResult"
        "400":
          description: No food detected or invalid image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ScanFoodRequest:
      type: object
      properties:
        imageBase64:
          type: string
          description: Base64 encoded image data
        mimeType:
          type: string
          description: MIME type of the image (e.g. image/jpeg)
      required:
        - imageBase64
        - mimeType
    FoodIngredient:
      type: object
      properties:
        name:
          type: string
        serving:
          type: string
        calories:
          type: number
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
      required:
        - name
        - serving
        - calories
        - protein
        - carbs
        - fat
    ScanFoodResult:
      type: object
      properties:
        isFood:
          type: boolean
        dishName:
          type: string
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/FoodIngredient"
        totals:
          type: object
          properties:
            calories:
              type: number
            protein:
              type: number
            carbs:
              type: number
            fat:
              type: number
          required:
            - calories
            - protein
            - carbs
            - fat
      required:
        - isFood
        - dishName
        - ingredients
        - totals
    FoodError:
      type: object
      properties:
        error:
          type: string
        isFood:
          type: boolean
      required:
        - error
        - isFood