# lib/api-zod/src/generated/types/scanFoodResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FoodIngredient } from "./foodIngredient";
import type { ScanFoodResultTotals } from "./scanFoodResultTotals";

export interface ScanFoodResult {
  isFood: boolean;
  dishName: string;
  ingredients: FoodIngredient[];
  totals: ScanFoodResultTotals;
}